move SliverPersistentHeader tests to the right file
